GH #339 Addressing HTTP2 issue on go1.13 above
