+++ .travis.yml
@@ -3,7 +3,7 @@ language: go
 sudo: false
 
 go: # use travis ci resource effectively, keep always latest 2 versions and tip :) 
-  - 1.12.x
+  - 1.14.x
   - 1.13.x
   - tip
